Feature/404 page (#6)
* Fixing MainMenu links to work on other pages

The main nav was only set up to smooth scroll for Projects and About me. Problem with this is if people want to go back the links never went anywhere. So have implemented logic to render either a smooth scrolling link or back to the homepage and anchor to the correct ID section

* Updating 404 page

* Removing duplicate About me link in MainMenu